# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ScoutOff API",
    "description": "Backend API for ScoutOff — a decentralised football scouting platform built on the Stellar blockchain (Soroban smart contracts).\n",
    "version": "1.0.0",
    "contact": {
      "name": "ScoutOff Engineering"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Stable API (version-less alias)"
    },
    {
      "url": "/api/v1",
      "description": "API v1"
    },
    {
      "url": "/api/v2",
      "description": "API v2 (includes SSE stream endpoint)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Long-lived scout API key (alternative to a Bearer JWT). Restricted keys may only perform operations covered by their granted scopes (#1019). Legacy keys without a scope list are unrestricted."
      }
    },
...
  },
  "paths": {
...
    "/events/stream": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Server-Sent Events endpoint.",
        "operationId": "getEventsStream",
        "description": "Server-Sent Events endpoint. Opens a long-lived HTTP connection and pushes\nrelevant contract events to the authenticated client as they are indexed.\n\nAuthentication: Bearer JWT (same as all other protected routes).\n\nQuery parameters (all optional, combinable):\n  - eventType  One or more event type names to subscribe to, comma-separated\n               (e.g. \"milestone_approved\" or\n               \"milestone_approved,scout_subscribed\"). When omitted the\n               client receives all event types that pass the\n               wallet-relevance filter. Unknown values are rejected with 400.\n  - playerId   Only deliver events whose payload contains this player identifier.\n               When omitted no additional player-level filtering is applied.\n\nFiltering: only events relevant to the authenticated wallet are sent (wallet\nisolation is always enforced regardless of query params). Player events are\nmatched through the authenticated wallet's player record, since their payload\ncarries a player ID rather than a wallet address. Optional query params add\nfurther narrowing on top.\n\nSSE event types sent:\n  - milestone_approved  (player: their own milestone approvals)\n  - scout_subscribed    (scout: their own subscription changes)\n  - contact_unlocked    (scout: their own contact unlocks)\n  - trial_offer_logged  (scout/player: trial offers involving them)\n  - player_registered   (player: their own registration)\n  - milestone_submitted (player/validator)\n  - fees_withdrawn      (admin)\n\nLive authorization enforcement (#1019):\n  - If the authenticated JWT is revoked (via POST /auth/logout or admin\n    token revocation) while the stream is open, the connection emits a\n    terminal `session_ended` event (reason \"token_revoked\") and closes;\n    no further protected events are delivered.\n  - If the wallet is blocklisted while the stream is open, the same\n    termination happens with reason \"wallet_blocklisted\".\n  - When the access JWT expires, the stream closes with reason \"token_expired\".\n  - Detection bound: immediate for revocations/blocklists processed in\n    this process; ≤ SSE_AUTH_SWEEP_INTERVAL_MS (default 30 s) for changes\n    persisted by another instance (one sweep query per process, never a\n    DB query per keep-alive tick).\n  - Blocklisted wallets cannot open a new connection (403).\n  - Concurrent streams are limited per wallet by\n    SSE_MAX_CONNECTIONS_PER_WALLET (default 5; 0 = unlimited).\n\nReconnection: initial `retry:` hint is sent on connect (configured via\nSSE_RETRY_MS, default 5000 ms + up to 20% random jitter) to prevent reconnect storms.\n\nKeep-alive: a `: ping` comment is sent every SSE_KEEPALIVE_INTERVAL_MS ms\n(default 15 s) to prevent idle-connection timeouts.",
        "responses": {
          "200": {
            "description": "text/event-stream — long-lived SSE connection"
          },
          "400": {
            "description": "{ success: false, error: string, code: string, validEventTypes: string[] } — unknown eventType"
          },
          "401": {
            "description": "{ success: false, error: string } — missing or invalid token"
          },
          "403": {
            "description": "{ success: false, error: string, code: 'WALLET_BLOCKLISTED' } — wallet is blocklisted"
          },
          "429": {
            "description": "{ success: false, error: string } — per-wallet connection limit reached"
          },
          "503": {
            "description": "{ success: false, error: string, code: 'SSE_CAPACITY' } — global connection limit reached (sets Retry-After)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/scouts/{wallet}/subscribe": {
      "post": {
        "tags": [
          "scouts"
        ],
        "summary": "Purchase a new scout subscription.",
        "operationId": "subscribe",
        "description": "Purchase a new scout subscription.",
        "parameters": [
          {
            "name": "wallet",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Scout's Stellar public key"
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Optional. Ensures safe retries: duplicate keys return\n  the cached response for 24 hours without triggering a new on-chain transaction."
          }
        ],
        "requestBody": {
          "required": true,
          "description": "{ tier: 'basic' | 'premium', duration: number (1–365 days) }",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tier": {
                    "type": "string",
                    "enum": [
                      "basic",
                      "premium"
                    ]
                  },
                  "duration": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 365
                  }
                },
                "required": [
                  "tier",
                  "duration"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "{ success: true, data: { transactionId, tier, expiresAt, status } }"
          },
          "400": {
            "description": "{ success: false, error: string } - Invalid tier or duration"
          },
          "402": {
            "description": "{ success: false, error: string } - Insufficient XLM balance"
          },
          "403": {
            "description": "{ success: false, error: string } - Scout role required or wallet mismatch"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": [
              "write:subscriptions"
            ]
          }
        ],
        "x-required-role": [
          "scout"
        ],
        "x-api-key-scope": "write:subscriptions",
        "x-owner-only": true
      },
      "put": {
        "tags": [
          "scouts"
        ],
        "summary": "Renew or create a subscription.",
        "operationId": "renewSubscription",
        "description": "Renew or create a subscription.\nIf an existing subscription exists, extends its expiry by `duration` days.\nIf no subscription exists, behaves like POST (creates a new one).",
        "parameters": [
          {
            "name": "wallet",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Scout's Stellar public key"
          }
        ],
        "requestBody": {
          "required": true,
          "description": "{ tier: 'basic' | 'premium', duration: number (1–365 days) }",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tier": {
                    "type": "string",
                    "enum": [
                      "basic",
                      "premium"
                    ]
                  },
                  "duration": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 365
                  }
                },
                "required": [
                  "tier",
                  "duration"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "{ success: true, data: { transactionId, tier, expiresAt, status } } - Renewal"
          },
          "201": {
            "description": "{ success: true, data: { transactionId, tier, expiresAt, status } } - New subscription"
          },
          "400": {
            "description": "{ success: false, error: string } - Invalid tier or duration"
          },
          "402": {
            "description": "{ success: false, error: string } - Insufficient XLM balance"
          },
          "403": {
            "description": "{ success: false, error: string } - Scout role required or wallet mismatch"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": [
              "write:subscriptions"
            ]
          }
        ],
        "x-required-role": [
          "scout"
        ],
        "x-api-key-scope": "write:subscriptions",
        "x-owner-only": true
      },
      "delete": {
        "tags": [
          "scouts"
        ],
        "summary": "Cancel an active subscription.",
        "operationId": "cancelSubscription",
        "description": "Cancel an active subscription. Records cancellation on-chain and locally.",
        "parameters": [
          {
            "name": "wallet",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Scout's Stellar public key"
          }
        ],
        "responses": {
          "200": {
            "description": "{ success: true, data: { transactionId, cancelledAt, wallet } }"
          },
          "403": {
            "description": "{ success: false, error: string } - Scout role required or wallet mismatch"
          },
          "404": {
            "description": "{ success: false, error: string } - No active subscription found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": [
              "write:subscriptions"
            ]
          }
        ],
        "x-required-role": [
          "scout"
        ],
        "x-api-key-scope": "write:subscriptions",
        "x-owner-only": true
      }
    },
...
  }
}